requirements.txt
# MQTT клиент
paho-mqtt>=1.6.1

# AWS SDK
boto3>=1.26.0

# Работа с Parquet файлами
pyarrow>=14.0.1

# Для конфигурации
python-dotenv>=1.0.0

# Асинхронные операции
aiohttp>=3.9.1

# Для работы с данными
pandas>=2.1.0

# Для разработки
setuptools>=68.0.0 